opentelemetry_sdk::export::metrics::aggregation
Trait Aggregation

pub trait Aggregation {
    // Required method
    fn kind(&self) -> &AggregationKind;
}

Available on crate feature metrics only.

An interface returned by an Aggregator containing an interval of metric data.

Required Methods

fn kind(&self) -> &AggregationKind

A short identifying string to identify the Aggregator that was used to produce the aggregation (e.g., AggregationKind::SUM).