Trait opentelemetry_sdk::export::metrics::InstrumentationLibraryReader

pub trait InstrumentationLibraryReader {
    // Required method
    fn try_for_each(
        &self,
        f: &mut dyn FnMut(&InstrumentationLibrary, &mut dyn Reader) -> Result<()>
    ) -> Result<()>;
}

Available on crate feature metrics only.

InstrumentationLibraryReader is an interface for exporters to iterate over one instrumentation library of metric data at a time.

Required Methods

fn try_for_each( &self, f: &mut dyn FnMut(&InstrumentationLibrary, &mut dyn Reader) -> Result<()> ) -> Result<()>

ForEach calls the passed function once per instrumentation library, allowing the caller to emit metrics grouped by the library that produced them.