opentelemetry_sdk::export::metrics
Trait LockedCheckpointer

pub trait LockedCheckpointer {
    // Required methods
    fn processor(&mut self) -> &mut dyn LockedProcessor;
    fn reader(&mut self) -> &mut dyn Reader;
    fn start_collection(&mut self);
    fn finish_collection(&mut self) -> Result<()>;
}

Available on crate feature metrics only.

The interface used by a controller to coordinate the processor with accumulator(s) and exporter(s).

The StartCollection() and FinishCollection() methods start and finish a collection interval. Controllers call the Accumulator(s) during collection to process Accumulations.

Required Methods

fn processor(&mut self) -> &mut dyn LockedProcessor

Processes metric data for export.

The process method is bracketed by start_collection and finish_collection calls.

fn reader(&mut self) -> &mut dyn Reader

Reader returns the current data set.

This may be called before and after collection. The implementation is required to return the same value throughout its lifetime.

fn start_collection(&mut self)

begins a collection interval.

fn finish_collection(&mut self) -> Result<()>

ends a collection interval.