opentelemetry_sdk

Module logs

OpenTelemetry Log SDK

Re-exports

pub use in_memory_exporter::InMemoryLogExporter;

pub use in_memory_exporter::InMemoryLogExporterBuilder;

Modules

in_memory_exporter

In-Memory log exporter for testing purpose.

Structs

BatchConfig

Batch log processor configuration. Use BatchConfigBuilder to configure your own instance of BatchConfig.

BatchConfigBuilder

A builder for creating BatchConfig instances.

BatchLogProcessor

The BatchLogProcessor collects finished logs in a buffer and exports them in batches to the configured LogExporter. This processor is ideal for high-throughput environments, as it minimizes the overhead of exporting logs individually. It uses a dedicated background thread to manage and export logs asynchronously, ensuring that the application’s main execution flow is not blocked.

BatchLogProcessorBuilder

A builder for creating BatchLogProcessor instances.

LogBatch

A batch of log records to be exported by a LogExporter.

LoggerProviderBuilder

Builder for provider attributes.

SdkLogRecord

LogRecord represents all data carried by a log record, and is provided to LogExporters as input.

SdkLogger

The object for emitting LogRecords.

SdkLoggerProvider

Handles the creation and coordination of Loggers.

SimpleLogProcessor

A LogProcessor designed for testing and debugging purpose, that immediately exports log records as they are emitted. Log records are exported synchronously in the same thread that emits the log record. When using this processor with the OTLP Exporter, the following exporter features are supported:

TraceContext

TraceContext stores the trace context for logs that have an associated span.

Traits

LogExporter

LogExporter defines the interface that log exporters should implement.

LogProcessor

The interface for plugging into a SdkLogger.