opentelemetry_sdk::logs

Enum LogError

#[non_exhaustive]
pub enum LogError {
    ExportFailed(Box<dyn ExportError>),
    ExportTimedOut(Duration),
    AlreadyShutdown(String),
    MutexPoisoned(String),
    Other(Box<dyn Error + Send + Sync + 'static>),
}

Errors returned by the log SDK.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ExportFailed(Box<dyn ExportError>)

Export failed with the error returned by the exporter.

ExportTimedOut(Duration)

Export failed to finish after certain period and processor stopped the export.

AlreadyShutdown(String)

Processor is already shutdown

MutexPoisoned(String)

Mutex lock poisoning

Other(Box<dyn Error + Send + Sync + 'static>)

Other errors propagated from log SDK that werenโt covered above.