opentelemetry_sdk
Module metrics

Available on crate feature metrics only.

The rust of the OpenTelemetry metrics SDK.

Configuration

The metrics SDK configuration is stored with each MeterProvider. Configuration for Resources, Views, and ManualReader or PeriodicReader instances can be specified.

Example

use opentelemetry_api::{
    metrics::{MeterProvider as _, Unit},
    KeyValue,
};
use opentelemetry_sdk::{metrics::MeterProvider, Resource};

// Generate SDK configuration, resource, views, etc
let resource = Resource::default(); // default attributes about the current process

// Create a meter provider with the desired config
let provider = MeterProvider::builder().with_resource(resource).build();

// Use the meter provider to create meter instances
let meter = provider.meter("my_app");

// Create instruments scoped to the meter
let counter = meter
    .u64_counter("power_consumption")
    .with_unit(Unit::new("kWh"))
    .init();

// use instruments to record measurements
counter.add(10, &[KeyValue::new("rate", "standard")]);

Re-exports

pub use pipeline::Pipeline;

Modules

data

Types for delivery of pre-aggregated metric time series data.

exporter

Interfaces for exporting metrics

reader

Interfaces for reading and producing metrics

Structs

Instrument

Describes properties an instrument is created with, also used for filtering in Views.

ManualReader

A simple MetricReader that allows an application to read metrics on demand.

ManualReaderBuilder

Configuration for a ManualReader

Meter

Handles the creation and coordination of all metric instruments.

MeterProvider

Handles the creation and coordination of Meters.

MeterProviderBuilder

Configuration options for a MeterProvider.

PeriodicReader

A MetricReader that continuously collects and exports metric data at a set interval.

PeriodicReaderBuilder

Configuration options for PeriodicReader.

Stream

Describes the stream of data an instrument produces.

Enums

Aggregation

The way recorded measurements are summarized.

InstrumentKind

The identifier of a group of instruments that all perform the same function.

Traits

View

Used to customize the metrics that are output by the SDK.

Functions

new_view

Creates a View that applies the Stream mask for all instruments that match criteria.