opentelemetry_sdk::metrics::aggregators
Struct HistogramAggregator

pub struct HistogramAggregator { /* private fields */ }

Available on crate feature metrics only.

This aggregator observes events and counts them in pre-determined buckets. It also calculates the sum and count of all events.