opentelemetry_sdk::metrics
Module data

Available on crate feature metrics only.

Types for delivery of pre-aggregated metric time series data.

Structs

DataPoint

DataPoint is a single data point in a time series.

Exemplar

A measurement sampled from a time series providing a typical example.

Gauge

A measurement of the current value of an instrument.

Histogram

Represents the histogram of all measurements of values from an instrument.

HistogramDataPoint

A single histogram data point in a time series.

Metric

A collection of one or more aggregated time series from an Instrument.

ResourceMetrics

A collection of ScopeMetrics and the associated Resource that created them.

ScopeMetrics

A collection of metrics produced by a meter.

Sum

Represents the sum of all measurements of values from an instrument.

Enums

Temporality

Defines the window that an aggregation was calculated over.

Traits

Aggregation

The store of data reported by an Instrument.