opentelemetry_sdk::metrics::data

Struct ExponentialHistogramDataPoint

pub struct ExponentialHistogramDataPoint<T> { /* private fields */ }

A single exponential histogram data point in a time series.

Implementations

impl<T> ExponentialHistogramDataPoint<T>

pub fn attributes(&self) -> impl Iterator<Item = &KeyValue>

Returns an iterator over the attributes in ExponentialHistogramDataPoint.

pub fn exemplars(&self) -> impl Iterator<Item = &Exemplar<T>>

Returns an iterator over the exemplars in ExponentialHistogramDataPoint.

pub fn count(&self) -> usize

Returns the number of updates this histogram has been calculated with.

pub fn scale(&self) -> i8

Returns the resolution of the histogram.

pub fn zero_count(&self) -> u64

Returns the number of values whose absolute value is less than or equal to zero_threshold.

pub fn positive_bucket(&self) -> &ExponentialBucket

Returns the range of positive value bucket counts.

pub fn negative_bucket(&self) -> &ExponentialBucket

Returns the range of negative value bucket counts.

pub fn zero_threshold(&self) -> f64

Returns the width of the zero region.

impl<T: Copy> ExponentialHistogramDataPoint<T>

pub fn min(&self) -> Option<T>

Returns the minimum value recorded.

pub fn max(&self) -> Option<T>

Returns the maximum value recorded.

pub fn sum(&self) -> T

Returns the sum of the values recorded.