opentelemetry_sdk::metrics
Module sdk_api

Available on crate feature metrics only.

SDK API

Structs

AtomicNumber

An atomic version of Number

Descriptor

Descriptor contains all the settings that describe an instrument, including its name, metric kind, number kind, and the configurable options.

Number

Number represents either an integral or a floating point value. It needs to be accompanied with a source of NumberKind that describes the actual type of the value stored within Number.

Enums

InstrumentKind

Kinds of OpenTelemetry metric instruments

NumberKind

A descriptor for the encoded data type of a Number

Traits

AsDynInstrumentCore

A utility extension to allow upcasting.

AsyncInstrumentCore

An implementation-level interface to an asynchronous instrument (e.g., Observable instruments).

InstrumentCore

A common interface for synchronous and asynchronous instruments.

MeterCore

The interface an SDK must implement to supply a Meter implementation.

SyncInstrumentCore

The implementation-level interface to a generic synchronous instrument (e.g., Histogram and Counter instruments).

Functions

wrap_meter_core

wraps impl to be a full implementation of a Meter.