opentelemetry_sdk

Module runtime

Provides an abstraction of several async runtimes

This allows OpenTelemetry to work with any current or future runtime. There is currently built-in implementation for Tokio.

Structs

Tokio

Runtime implementation, which works with Tokio’s multi thread runtime.

TokioCurrentThread

Runtime implementation, which works with Tokio’s current thread runtime.

Enums

TrySendError

Error returned by a TrySend implementation.

Traits

Runtime

A runtime is an abstraction of an async runtime like Tokio. It allows OpenTelemetry to work with any current and hopefully future runtime implementations.

RuntimeChannel

RuntimeChannel is an extension to Runtime. Currently, it provides a channel that is used by the log and span batch processors.

TrySend

TrySend is an abstraction of Sender that is capable of sending messages through a reference.