opentelemetry_sdk::trace

Struct Config

#[non_exhaustive]
pub struct Config {
    pub sampler: Box<dyn ShouldSample>,
    pub id_generator: Box<dyn IdGenerator>,
    pub span_limits: SpanLimits,
    pub resource: Cow<'static, Resource>,
}

Tracer configuration

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

sampler: Box<dyn ShouldSample>

The sampler that the sdk should use

id_generator: Box<dyn IdGenerator>

The id generator that the sdk should use

span_limits: SpanLimits

span limits

resource: Cow<'static, Resource>

Contains attributes representing an entity that produces telemetry.

Implementations

impl Config

pub fn with_sampler<T: ShouldSample + 'static>(self, sampler: T) -> Self

👎Deprecated since 0.27.1: Config is becoming private. Please use Builder::with_sampler(…) instead.

Specify the sampler to be used.

pub fn with_id_generator<T: IdGenerator + 'static>( self, id_generator: T, ) -> Self

👎Deprecated since 0.27.1: Config is becoming private. Please use Builder::with_id_generator(…) instead.

Specify the id generator to be used.

pub fn with_max_events_per_span(self, max_events: u32) -> Self

👎Deprecated since 0.27.1: Config is becoming private. Please use Builder::with_max_events_per_span(…) instead.

Specify the maximum number of events that can be recorded per span.

pub fn with_max_attributes_per_span(self, max_attributes: u32) -> Self

👎Deprecated since 0.27.1: Config is becoming private. Please use Builder::with_max_attributes_per_span(…) instead.

Specify the maximum number of attributes that can be recorded per span.

pub fn with_max_links_per_span(self, max_links: u32) -> Self

👎Deprecated since 0.27.1: Config is becoming private. Please use Builder::with_max_links_per_span(…) instead.

Specify the maximum number of links that can be recorded per span.

pub fn with_max_attributes_per_event(self, max_attributes: u32) -> Self

👎Deprecated since 0.27.1: Config is becoming private. Please use Builder::with_max_attributes_per_event(…) instead.

Specify the maximum number of attributes one event can have.

pub fn with_max_attributes_per_link(self, max_attributes: u32) -> Self

👎Deprecated since 0.27.1: Config is becoming private. Please use Builder::with_max_attributes_per_link(…) instead.

Specify the maximum number of attributes one link can have.

pub fn with_span_limits(self, span_limits: SpanLimits) -> Self

👎Deprecated since 0.27.1: Config is becoming private. Please use Builder::with_span_limits(…) instead.

Specify all limit via the span_limits

pub fn with_resource(self, resource: Resource) -> Self

👎Deprecated since 0.27.1: Config is becoming private. Please use Builder::with_resource(…) instead.

Specify the attributes representing the entity that produces telemetry