Struct opentelemetry_sdk::trace::EvictedQueue

pub struct EvictedQueue<T> { /* private fields */ }

This queue maintains an ordered list of elements, and a count of dropped elements. Elements are removed from the queue in a first in first out fashion.

Implementations

impl<T> EvictedQueue<T>

pub fn new(max_len: u32) -> Self

Create a new EvictedQueue with a given max length.

pub fn append_vec(&mut self, other: &mut Vec<T>)

Moves all the elements of other into self, leaving other empty.

pub fn is_empty(&self) -> bool

Returns true if the EvictedQueue is empty.

pub fn iter(&self) -> Iter<'_, T>

Returns a front-to-back iterator.

pub fn len(&self) -> usize

Returns the number of elements in the EvictedQueue.

pub fn dropped_count(&self) -> u32

Count of dropped attributes