Crate opentelemetry_sdk

OpenTelemetry SDK

This SDK provides an opinionated reference implementation of the OpenTelemetry API. The SDK implements the specifics of deciding which data to collect through Samplers, and facilitates the delivery of telemetry data to storage systems through Exporters. These can be configured on Tracer and Meter creation.

Modules

export

Metrics Export

metrics

OpenTelemetry Metrics SDK

propagation

OpenTelemetry Propagators

resource

Representations of entities producing telemetry.

runtime

Provides an abstraction of several async runtimes

trace

OpenTelemetry Trace SDK

Structs

InstrumentationLibrary

Information about a library or crate providing instrumentation.

Resource

An immutable representation of the entity producing telemetry as attributes.