Crate opentelemetry_sdk

Implements the SDK component of OpenTelemetry.

Supported Rust Versions

Getting Started

use opentelemetry::{global, trace::{Tracer, TracerProvider as _}};
use opentelemetry_sdk::trace::TracerProvider;

fn main() {
    // Choose an exporter like `opentelemetry_stdout::SpanExporter`
    let exporter = new_exporter();

    // Create a new trace pipeline that prints to stdout
    let provider = TracerProvider::builder()
        .with_simple_exporter(exporter)
        .build();
    let tracer = provider.tracer("readme_example");

    tracer.in_span("doing_work", |cx| {
        // Traced app logic here...
    });

    // Shutdown trace pipeline
    global::shutdown_tracer_provider();
}

See the examples directory for different integration patterns.

See the API trace module docs for more information on creating and managing spans.

Metrics (Alpha)

Note: the metrics implementation is still in progress and subject to major changes.

Creating instruments and recording measurements

use opentelemetry::{global, KeyValue};

// get a meter from a provider
let meter = global::meter("my_service");

// create an instrument
let counter = meter.u64_counter("my_counter").build();

// record a measurement
counter.add(1, &[KeyValue::new("http.client_ip", "83.164.160.102")]);

See the examples directory for different integration patterns.

See the API metrics module docs for more information on creating and managing instruments.

Crate Feature Flags

The following feature flags can used to control the telemetry signals to use:

trace: Includes the trace SDK (enabled by default).

metrics: Includes the metrics SDK.

logs: Includes the logs SDK.

For trace the following feature flags are available:

jaeger_remote_sampler: Enables the Jaeger remote sampler.

For logs the following feature flags are available:

spec_unstable_logs_enabled: control the log level

Support for recording and exporting telemetry asynchronously and perform metrics aggregation can be added via the following flags:

rt-tokio: Spawn telemetry tasks using tokio’s multi-thread runtime.

rt-tokio-current-thread: Spawn telemetry tasks on a separate runtime so that the main runtime won’t be blocked.

rt-async-std: Spawn telemetry tasks using async-std’s runtime.

Modules

error

Wrapper for error from trace, logs and metrics part of open telemetry.

export

Telemetry Export

logs

OpenTelemetry Log SDK

metrics

The crust of the OpenTelemetry metrics SDK.

propagation

OpenTelemetry Propagators

resource

Representations of entities producing telemetry.

runtime

Provides an abstraction of several async runtimes

trace

OpenTelemetry Trace SDK

Structs

Resource

An immutable representation of the entity producing telemetry as attributes. Utilizes Arc for efficient sharing and cloning.