opentelemetry_sdk::metrics

Module data

Types for delivery of pre-aggregated metric time series data.

Structs

Exemplar

A measurement sampled from a time series providing a typical example.

ExponentialBucket

A set of bucket counts, encoded in a contiguous array of counts.

ExponentialHistogram

The histogram of all measurements of values from an instrument.

ExponentialHistogramDataPoint

A single exponential histogram data point in a time series.

Gauge

A measurement of the current value of an instrument.

GaugeDataPoint

DataPoint is a single data point in a time series.

Histogram

Represents the histogram of all measurements of values from an instrument.

HistogramDataPoint

A single histogram data point in a time series.

Metric

A collection of one or more aggregated time series from an Instrument.

ResourceMetrics

A collection of ScopeMetrics and the associated Resource that created them.

ScopeMetrics

A collection of metrics produced by a meter.

Sum

Represents the sum of all measurements of values from an instrument.

SumDataPoint

DataPoint is a single data point in a time series.

Enums

AggregatedMetrics

Aggregated metrics data from an instrument

MetricData

Metric data for all types