opentelemetry_sdk

Module trace

OpenTelemetry Trace SDK

The tracing SDK consist of a few main structs:

The Tracer struct which performs all tracing operations.

The Span struct with is a mutable object storing information about the current operation execution.

The TracerProvider struct which configures and produces Tracers.

Structs

BatchConfig

Batch span processor configuration. Use BatchConfigBuilder to configure your own instance of BatchConfig.

BatchConfigBuilder

A builder for creating BatchConfig instances.

BatchSpanProcessor

A SpanProcessor that asynchronously buffers finished spans and reports them at a preconfigured interval.

BatchSpanProcessorBuilder

A builder for creating BatchSpanProcessor instances.

Builder

Builder for provider attributes.

Config

Tracer configuration

RandomIdGenerator

Default IdGenerator implementation.

SimpleSpanProcessor

A SpanProcessor that passes finished spans to the configured SpanExporter, as soon as they are finished, without any batching. This is typically useful for debugging and testing. For scenarios requiring higher performance/throughput, consider using BatchSpanProcessor.

Span

Single operation within a trace.

SpanEvents

Stores span events along with dropped count.

SpanLimits

Span limit configuration to keep attributes, events and links to a span in a reasonable number.

SpanLinks

Stores span links along with dropped count.

Tracer

Tracer implementation to create and manage spans

TracerProvider

Creator and registry of named Tracer instances.

Enums

Sampler

Default Sampling options

Traits

IdGenerator

Interface for generating IDs

ShouldSample

The ShouldSample interface allows implementations to provide samplers which will return a sampling SamplingResult based on information that is typically available just before the Span was created.

SpanProcessor

SpanProcessor is an interface which allows hooks for span start and end method invocations. The span processors are invoked only when is_recording is true.

Functions

configDeprecated

Default trace configuration